examples/python/eds_server_example/requirements.txt
asyncio==3.4.3
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
nats-py==2.2.0
Pygments==2.14.0
rich==12.6.0
shellingham==1.5.0.post1
typer==0.7.0
